+++ src/main/java/com/dissi/adventofcode/version2022/day09/Ekans.java
@@ -1,6 +1,6 @@
 package com.dissi.adventofcode.version2022.day09;
 
-import static java.lang.Integer.signum;
+import static java.lang.Long.signum;
 
 import com.dissi.adventofcode.helpers.Direction;
 import com.dissi.adventofcode.helpers.Position;
@@ -15,8 +15,8 @@ public class Ekans {
 
     public static Position keepUpTail(Position head, Position tail) {
         Position delta = head.substract(tail);
-        int dx = delta.getX();
-        int dy = delta.getY();
+        long dx = delta.getX();
+        long dy = delta.getY();
         if (Math.abs(dy) == 2 || Math.abs(dx) == 2) {
             return tail.add(signum(dx), signum(dy));
         }
